Update type and location of initialization for z.

+++ src/gource.h
@@ -176,7 +176,7 @@ class Gource : public SDLApp {
     Bounds2D active_user_bounds;
 
     int commitqueue_max_size;
-    int starting_z;
+    float starting_z;
 
     std::deque<RCommit> commitqueue;
     std::map<std::string, RUser*> users;
